hosts/lia/sshfwd/mkFwdSrv: bump RestartSec
The default limit is to allow 5 restarts in a 10sec period. If a service goes
over that threshold due to the Restart= option, it will not attempt to restart
any further.

The rates are configured with the StartLimitIntervalSec= and StartLimitBurst=
options and the Restart= option controls when SystemD tries to restart a
service.