kay/network: change lan port
the usb ethernet device don't seems to work on the usb2.0 ports anymore.
the last working commit was cd942d253bda8f511fdb921ea29f69f382a9368e.
it's probably due to external nixpkgs changes, so make sure to check all
the flake bumps in between if the need arises in the future.